Offsets to remove diagonal artifacts
Previously different levels of the fractal were simply scaled versions
of each other with respect to the origin. This lead to lines streaking
out from the origin where patterns on different levels would align.

To counter this, each level is created with a different Noise object and
each has it's own index offset.