Removed index offsets in favour of unique gradients per noise object
In order to fix diagonal artifacts in the fractal results, an index
offset was introduced in order to have each level indexing a different
place in the same block of gradients.

However we were still storing the gradients on a per object basis, so
this has been removed in favour of having different gradients per noise
object so that the levels in the fractal do not coincide.